# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CatImpact": {
        "description": "A projected category total for the optimal lineup (display string + trend).",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "proj": {
            "title": "Proj",
            "type": "string"
          },
          "trend": {
            "default": "flat",
            "title": "Trend",
            "type": "string"
          }
        },
        "required": [
          "key",
          "proj"
        ],
        "title": "CatImpact",
        "type": "object"
      },
      "CategoryImpact": {
        "properties": {
          "cat": {
            "title": "Cat",
            "type": "string"
          },
          "delta": {
            "title": "Delta",
            "type": "number"
          }
        },
        "required": [
          "cat",
          "delta"
        ],
        "title": "CategoryImpact",
        "type": "object"
      },
...
      "DailyMeta": {
        "description": "Day-level context for the daily optimizer (populated only when mode='daily').",
        "properties": {
          "ip_pace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IpPace"
              },
              {
                "type": "null"
              }
            ]
          },
          "losing": {
            "items": {
              "type": "string"
            },
            "title": "Losing",
            "type": "array"
          },
          "rate_modes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Rate Modes",
            "type": "object"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "swaps": {
            "items": {
              "$ref": "#/components/schemas/Swap"
            },
            "title": "Swaps",
            "type": "array"
          },
          "tied": {
            "items": {
              "type": "string"
            },
            "title": "Tied",
            "type": "array"
          },
          "urgency": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Urgency",
            "type": "object"
          },
          "winning": {
            "items": {
              "type": "string"
            },
            "title": "Winning",
            "type": "array"
          }
        },
        "title": "DailyMeta",
        "type": "object"
      },
...
      "DraftClock": {
        "properties": {
          "current_pick": {
            "title": "Current Pick",
            "type": "integer"
          },
          "is_user_turn": {
            "title": "Is User Turn",
            "type": "boolean"
          },
          "pick_in_round": {
            "title": "Pick In Round",
            "type": "integer"
          },
          "picking_team_index": {
            "title": "Picking Team Index",
            "type": "integer"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          }
        },
        "required": [
          "current_pick",
          "round",
          "pick_in_round",
          "picking_team_index",
          "is_user_turn"
        ],
        "title": "DraftClock",
        "type": "object"
      },
      "DraftConfig": {
        "properties": {
          "num_rounds": {
            "default": 23,
            "title": "Num Rounds",
            "type": "integer"
          },
          "num_teams": {
            "default": 12,
            "title": "Num Teams",
            "type": "integer"
          },
          "roster_config": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roster Config"
          },
          "user_team_index": {
            "default": 0,
            "title": "User Team Index",
            "type": "integer"
          }
        },
        "title": "DraftConfig",
        "type": "object"
      },
      "DraftPick": {
        "properties": {
          "pick": {
            "title": "Pick",
            "type": "integer"
          },
          "player_id": {
            "title": "Player Id",
            "type": "integer"
          },
          "player_name": {
            "title": "Player Name",
            "type": "string"
          },
          "positions": {
            "title": "Positions",
            "type": "string"
          },
          "team_index": {
            "title": "Team Index",
            "type": "integer"
          }
        },
        "required": [
          "pick",
          "team_index",
          "player_id",
          "player_name",
          "positions"
        ],
        "title": "DraftPick",
        "type": "object"
      },
      "DraftRecommendRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/DraftConfig",
            "default": {
              "num_rounds": 23,
              "num_teams": 12,
              "user_team_index": 0
            }
          },
          "n_simulations": {
            "default": 300,
            "title": "N Simulations",
            "type": "integer"
          },
          "pick_log": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DraftPick"
            },
            "title": "Pick Log",
            "type": "array"
          },
          "top_n": {
            "default": 8,
            "title": "Top N",
            "type": "integer"
          }
        },
        "title": "DraftRecommendRequest",
        "type": "object"
      },
      "DraftRecommendResponse": {
        "properties": {
          "clock": {
            "$ref": "#/components/schemas/DraftClock"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/DraftRecommendation"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "clock",
          "recommendations"
        ],
        "title": "DraftRecommendResponse",
        "type": "object"
      },
      "DraftRecommendation": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerRef"
          },
          "projected_sgp": {
            "title": "Projected Sgp",
            "type": "number"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          }
        },
        "required": [
          "player",
          "rank",
          "score",
          "projected_sgp"
        ],
        "title": "DraftRecommendation",
        "type": "object"
      },
      "DraftSimulatePicksRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/DraftConfig",
            "default": {
              "num_rounds": 23,
              "num_teams": 12,
              "user_team_index": 0
            }
          },
          "pick_log": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DraftPick"
            },
            "title": "Pick Log",
            "type": "array"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "title": "DraftSimulatePicksRequest",
        "type": "object"
      },
      "DraftSimulatePicksResponse": {
        "properties": {
          "clock": {
            "$ref": "#/components/schemas/DraftClock"
          },
          "picks": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DraftPick"
            },
            "title": "Picks",
            "type": "array"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "clock"
        ],
        "title": "DraftSimulatePicksResponse",
        "type": "object"
      },
...
      "GradeRange": {
        "properties": {
          "center": {
            "title": "Center",
            "type": "number"
          },
          "grade": {
            "title": "Grade",
            "type": "string"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          }
        },
        "required": [
          "grade",
          "low",
          "center",
          "high"
        ],
        "title": "GradeRange",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IpPace": {
        "description": "Weekly innings-pitched pacing for the daily view (from ip_tracker).",
        "properties": {
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "pace_pct": {
            "title": "Pace Pct",
            "type": "integer"
          },
          "projected": {
            "title": "Projected",
            "type": "number"
          },
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "number"
          }
        },
        "required": [
          "projected",
          "target",
          "pace_pct"
        ],
        "title": "IpPace",
        "type": "object"
      },
...
      "LineupOptimizeRequest": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "mode": {
            "default": "standard",
            "title": "Mode",
            "type": "string"
          },
          "scope": {
            "default": "rest_of_season",
            "title": "Scope",
            "type": "string"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          }
        },
        "required": [
          "team_name"
        ],
        "title": "LineupOptimizeRequest",
        "type": "object"
      },
      "LineupOptimizeResponse": {
        "properties": {
          "bench": {
            "items": {
              "$ref": "#/components/schemas/LineupSlot"
            },
            "title": "Bench",
            "type": "array"
          },
          "daily": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DailyMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "impact": {
            "items": {
              "$ref": "#/components/schemas/CatImpact"
            },
            "title": "Impact",
            "type": "array"
          },
          "mode": {
            "default": "standard",
            "title": "Mode",
            "type": "string"
          },
          "optimal": {
            "default": false,
            "title": "Optimal",
            "type": "boolean"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/LineupSlot"
            },
            "title": "Slots",
            "type": "array"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          }
        },
        "required": [
          "team_name",
          "date",
          "slots"
        ],
        "title": "LineupOptimizeResponse",
        "type": "object"
      },
...
      "LineupSlot": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "current_slot": {
            "default": "",
            "title": "Current Slot",
            "type": "string"
          },
          "forced_start": {
            "default": false,
            "title": "Forced Start",
            "type": "boolean"
          },
          "matchup": {
            "default": "",
            "title": "Matchup",
            "type": "string"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerRef"
          },
          "projected": {
            "default": 0.0,
            "title": "Projected",
            "type": "number"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "slot": {
            "title": "Slot",
            "type": "string"
          },
          "status": {
            "default": "start",
            "title": "Status",
            "type": "string"
          },
          "value": {
            "default": 0.0,
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "slot",
          "player",
          "action"
        ],
        "title": "LineupSlot",
        "type": "object"
      },
...
      "PlayerRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mlb_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlb Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "positions": {
            "title": "Positions",
            "type": "string"
          },
          "team_abbr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Abbr"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "yahoo_player_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yahoo Player Key"
          }
        },
        "required": [
          "id",
          "name",
          "positions"
        ],
        "title": "PlayerRef",
        "type": "object"
      },
...
      "Swap": {
        "description": "A recommended move: start this currently-benched player in `slot`.",
        "properties": {
          "player": {
            "$ref": "#/components/schemas/PlayerRef"
          },
          "slot": {
            "title": "Slot",
            "type": "string"
          },
          "value": {
            "default": 0.0,
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "player",
          "slot"
        ],
        "title": "Swap",
        "type": "object"
      },
...
      "TradeEvaluateRequest": {
        "properties": {
          "enable_mc": {
            "default": false,
            "title": "Enable Mc",
            "type": "boolean"
          },
          "giving_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Giving Ids",
            "type": "array"
          },
          "receiving_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Receiving Ids",
            "type": "array"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          }
        },
        "required": [
          "team_name",
          "giving_ids",
          "receiving_ids"
        ],
        "title": "TradeEvaluateRequest",
        "type": "object"
      },
      "TradeEvaluationResponse": {
        "properties": {
          "category_impacts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CategoryImpact"
            },
            "title": "Category Impacts",
            "type": "array"
          },
          "confidence_pct": {
            "default": 0.0,
            "title": "Confidence Pct",
            "type": "number"
          },
          "delta_champ_prob": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Champ Prob"
          },
          "delta_playoff_prob": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Playoff Prob"
          },
          "giving": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlayerRef"
            },
            "title": "Giving",
            "type": "array"
          },
          "grade": {
            "default": "",
            "title": "Grade",
            "type": "string"
          },
          "grade_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GradeRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "mc_enabled": {
            "default": false,
            "title": "Mc Enabled",
            "type": "boolean"
          },
          "receiving": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlayerRef"
            },
            "title": "Receiving",
            "type": "array"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "surplus_sgp": {
            "default": 0.0,
            "title": "Surplus Sgp",
            "type": "number"
          },
          "verdict": {
            "default": "",
            "title": "Verdict",
            "type": "string"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "title": "TradeEvaluationResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "HEATER API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/draft/recommend": {
      "post": {
        "operationId": "draft_recommend_api_draft_recommend_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRecommendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftRecommendResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required (when billing is enabled)."
          },
          "402": {
            "description": "Pro subscription required (when billing is enabled)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Recommend",
        "tags": [
          "draft"
        ]
      }
    },
    "/api/draft/simulate-picks": {
      "post": {
        "operationId": "draft_simulate_picks_api_draft_simulate_picks_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftSimulatePicksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftSimulatePicksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required (when billing is enabled)."
          },
          "402": {
            "description": "Pro subscription required (when billing is enabled)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Simulate Picks",
        "tags": [
          "draft"
        ]
      }
    },
...
    "/api/lineup/optimize": {
      "post": {
        "operationId": "optimize_lineup_api_lineup_optimize_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineupOptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineupOptimizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required (when billing is enabled)."
          },
          "402": {
            "description": "Pro subscription required (when billing is enabled)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Optimize Lineup",
        "tags": [
          "lineup"
        ]
      }
    },
...
    "/api/trade/evaluate": {
      "post": {
        "operationId": "evaluate_trade_endpoint_api_trade_evaluate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradeEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeEvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required (when billing is enabled)."
          },
          "402": {
            "description": "Pro subscription required (when billing is enabled)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Trade Endpoint",
        "tags": [
          "trade"
        ]
      }
    },
...
  }
}